add getter for "registerCall" to PersistentWorkflow.
The registerCall member is used by the Processor. Under special circumstances a copper user might want to create
a custom Processor implementation. Because the registerCall member is package protected, the user is forced to put the custom processor
implementation in the "org.copperengine.core.persistent" package. Which is quite annoying.